requirements.txt
discord.py
openai
mistralai
python-dotenv
google-search-results
pydantic
aiohttp
sqlalchemy
psycopg2
beautifulsoup4
Pillow
requests
urllib3
fastapi
uvicorn
